[Refactor]폴더명 수정 및 이에 따른 경로 리팩토링
